Move spinner() and retry() to goods from experimental
