+++ test-d/goods.test-d.ts
@@ -12,11 +12,17 @@
 // See the License for the specific language governing permissions and
 // limitations under the License.
 
+import { expectType } from 'tsd'
 import { $ } from '../src/core.js'
-import { echo, sleep } from '../src/goods.js'
+import { echo, sleep, spinner, retry, expBackoff } from '../src/goods.js'
 
 echo`Date is ${await $`date`}`
 echo('Hello, world!')
 
 await sleep('1s')
 await sleep(1000)
+
+expectType<'foo'>(await spinner(() => 'foo' as 'foo'))
+expectType<'bar'>(await spinner('title', () => 'bar' as 'bar'))
+expectType<'foo'>(await retry(0, () => 'foo' as 'foo'))
+expectType<Generator<number, void, unknown>>(expBackoff())
